#279 add correct validations for submit payment details
